Remove duplicated joint logic
This patch removes some unnecessary code which duplicates the logic for
mapping joint type to constrained dofs. Also changes the chain of
`if-end` to `if-else-end` since there should never be more than one
branch matching the condition. This patch is a result of finding a
off-by-one error in `count` (which should have started at 0 and not 1).